requirements.txt
SpeechRecognition
joblib
fastapi[standard]
uvicorn
azure.storage.blob
soundfile
numpy
